+++ src/__tests__/configuration.test.ts
@@ -1,0 +1,5 @@
+describe('use-http configuration', () => {
+  it('should create a default configuration with no arguments', () => {
+    
+  });
+});
